exam service validation added

+++ src/electron/services/exam.service.ts
@@ -1,13 +1,18 @@
 import { IExamBase } from "@/shared/interfaces/index.js";
 import { ExamRepository } from "../repositories/exam.repository.js";
+import { validateExam } from "../validation/index.js";
 
 export class ExamService {
   constructor(private readonly examRepository: ExamRepository) {}
 
   createExam(exam: IExamBase) {
+    if (exam.status === "final") {
+      validateExam(exam);
+    }
     return this.examRepository.createExam(exam);
   }
   findExamById(examId: number) {
+    if (!examId) throw new Error("Exam Id is required to retrieve exam");
     return this.examRepository.findExamById(examId);
   }
   findAllExams() {
@@ -18,6 +23,7 @@ export class ExamService {
   }
 
   deleteExam(examId: number) {
+    if (!examId) throw new Error("Exam Id is required to delete exam");
     return this.examRepository.deleteExam(examId);
   }
 }
